+++ integration-tests/cli/extensions-install.test.ts
@@ -5,7 +5,7 @@
  */
 
 import { expect, test } from 'vitest';
-import { TestRig } from './test-helper.js';
+import { TestRig } from '../test-helper.js';
 import { writeFileSync } from 'node:fs';
 import { join } from 'node:path';
