+++ integration-tests/cli/telemetry.test.ts
@@ -5,7 +5,7 @@
  */
 
 import { describe, it, expect } from 'vitest';
-import { TestRig } from './test-helper.js';
+import { TestRig } from '../test-helper.js';
 
 describe('telemetry', () => {
   it('should emit a metric and a log event', async () => {
