+++ integration-tests/interactive/mixed-input-crash.test.ts
@@ -5,7 +5,7 @@
  */
 
 import { describe, it, expect } from 'vitest';
-import { TestRig } from './test-helper.js';
+import { TestRig } from '../test-helper.js';
 
 describe('mixed input crash prevention', () => {
   it('should not crash when using mixed prompt inputs', async () => {
